Remove SignUpController and SignUpService, move methods to StaffController and StaffService.
